Move bias tests to separate class
